% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{compute_linear_trend_ts}
\alias{compute_linear_trend_ts}
\alias{compute_linear_trend}
\title{Calculate linear trend coefficients}
\usage{
compute_linear_trend_ts(ts)

compute_linear_trend(dataset, ...)
}
\arguments{
\item{dataset}{the dataset}
}
\value{
A tibble, containing the `results`, dataset `metadata`, name of the
  input dataset in `dataset`, name of the `method` applied to each time
  series (in this case, "compute_linear_trend_ts"), and additional `args`.
}
\description{
`compute_linear_trend_ts()` fits a simple linear model to its
  only input, a time series (represented by a numeric vector of values), and
  returns the coefficients

`compute_linear_trend()` applies `compute_linear_trend_ts()` to
  each abundance time series in its input data.
}
\details{
In the implementation of `compute_linear_trend()`, we use the `MATSS`
  function \code{\link[MATSS]{analysis_wrapper}} to construct the version of
  `compute_linear_trend_ts()` that can be applied to a whole dataset.
}
\examples{
\dontrun{
  sgs_data <- MATSS::get_sgs_data()
  compute_linear_trend(sgs_data)
}

}
