% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{compute_simpson_index}
\alias{compute_simpson_index}
\title{Calculate Simpson's diversity index}
\usage{
compute_simpson_index(dat)
}
\arguments{
\item{dat}{the dataset}
}
\value{
A vector of the diversity index calculations, one per time point.
}
\description{
Calculate Simpson's diversity index
}
\details{
There are multiple forms of Simpson's index. Here, we compute
  \eqn{1-D}, where \eqn{D_t = sum (p_{t,i})^2}, with \eqn{p_{t,i}} being the
  proportional abundance of species \eqn{i} in time index \eqn{t}.

  This calculation produces a vector of values, one for each time point.
}
\examples{
\dontrun{
  sgs_data <- MATSS::get_sgs_data()
  compute_simpson_index(sgs_data)
}

}
